test: 사용자와 답변이 다른 User인 경우 삭제할때 에러 테스트 코드 추가

+++ src/main/java/qna/domain/Answer.java
@@ -1,5 +1,6 @@
 package qna.domain;
 
+import qna.CannotDeleteException;
 import qna.NotFoundException;
 import qna.UnAuthorizedException;
 
@@ -107,7 +108,10 @@ public class Answer extends BaseTimeEntity {
                 '}';
     }
 
-    public DeleteHistory delete() {
+    public DeleteHistory delete(User loginUser) throws CannotDeleteException {
+        if (!isOwner(loginUser)) {
+            throw new CannotDeleteException("다른 사람이 쓴 답변이 있어 삭제할 수 없습니다.");
+        }
         deleted = true;
         return new DeleteHistory(ContentType.ANSWER, id, writer);
     }
